+++ tests/load_dumped_file_test.py
@@ -5,6 +5,7 @@
 import unittest
 
 from sklearn.svm import SVR
+
 from src.load_dumped_file import load_dumped_file
 from src.make_directory import make_directory
 from src.read_in import read_in
